refactor: rename @epicenter/server to @epicenter/server-elysia
Make the framework dependency explicit in the package name ahead of
adding @epicenter/server-cloudflare (Hono + Durable Objects) in Phase 2.
The rename clarifies the architecture: sync-core is pure, server-elysia
is the Elysia adapter layer, and each server-* is a deployment target.

- git mv packages/server → packages/server-elysia
- Update package name, dependencies, imports, JSDoc, READMEs, and spec
- Regenerate bun.lock
- All tests pass (14 server-elysia + 35 server-local)
