+++ packages/server-elysia/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@epicenter/server",
+	"name": "@epicenter/server-elysia",
 	"version": "0.0.1",
 	"main": "./src/index.ts",
 	"types": "./src/index.ts",
