+++ packages/server-elysia/specs/20260220T223633 collapse-sync-routes.md
