+++ packages/server-elysia/src/auth.ts
