+++ packages/server-elysia/src/discovery/awareness.ts
