+++ packages/server-local/package.json
@@ -15,7 +15,7 @@
 		"typecheck": "tsc --noEmit"
 	},
 	"dependencies": {
-		"@epicenter/server": "workspace:*",
+		"@epicenter/server-elysia": "workspace:*",
 		"@elysiajs/cors": "^1.4.1",
 		"@elysiajs/openapi": "^1.4.11",
 		"elysia": "^1.2.25",
