+++ packages/server-remote/package.json
@@ -16,7 +16,7 @@
 		"typecheck": "tsc --noEmit"
 	},
 	"dependencies": {
-		"@epicenter/server": "workspace:*",
+		"@epicenter/server-elysia": "workspace:*",
 		"@elysiajs/openapi": "^1.4.11",
 		"@tanstack/ai": "^0.5.1",
 		"@tanstack/ai-anthropic": "^0.5.0",
